string class

+++ Kernel/scheduler.h
@@ -1,6 +1,6 @@
 #pragma once
 
-void request_new_process(size_t range, void (*entry_function)());
+void request_new_process(uint64_t range, void (*entry_function)());
 
 void testl();
